+++ mdm/mdm/marshal_json.go
@@ -1,0 +1,177 @@
+package mdm
+
+import (
+	"encoding/json"
+	"fmt"
+)
+
+func (c *Command) MarshalJSON() ([]byte, error) {
+	switch c.RequestType {
+	case "ProfileList",
+		"ProvisioningProfileList",
+		"CertificateList",
+		"SecurityInfo",
+		"RestartDevice",
+		"ShutDownDevice",
+		"StopMirroring",
+		"ClearRestrictionsPassword",
+		"UserList",
+		"LogOutUser",
+		"PlayLostModeSound",
+		"DisableLostMode",
+		"DeviceLocation",
+		"TODO_remove":
+		var x = struct {
+			RequestType string `json:"request_type"`
+		}{
+			RequestType: c.RequestType,
+		}
+		return json.Marshal(&x)
+	case "InstallProfile":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*InstallProfile
+		}{
+			RequestType:    c.RequestType,
+			InstallProfile: c.InstallProfile,
+		}
+		return json.Marshal(&x)
+	case "RemoveProfile":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*RemoveProfile
+		}{
+			RequestType:   c.RequestType,
+			RemoveProfile: c.RemoveProfile,
+		}
+		return json.Marshal(&x)
+	case "InstallProvisioningProfile":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*InstallProvisioningProfile
+		}{
+			RequestType:                c.RequestType,
+			InstallProvisioningProfile: c.InstallProvisioningProfile,
+		}
+		return json.Marshal(&x)
+	case "RemoveProvisioningProfile":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*RemoveProvisioningProfile
+		}{
+			RequestType:               c.RequestType,
+			RemoveProvisioningProfile: c.RemoveProvisioningProfile,
+		}
+		return json.Marshal(&x)
+	case "InstalledApplicationList":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*InstalledApplicationList
+		}{
+			RequestType:              c.RequestType,
+			InstalledApplicationList: c.InstalledApplicationList,
+		}
+		return json.Marshal(&x)
+	case "DeviceInformation":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*DeviceInformation
+		}{
+			RequestType:       c.RequestType,
+			DeviceInformation: c.DeviceInformation,
+		}
+		return json.Marshal(&x)
+	case "DeviceLock":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*DeviceLock
+		}{
+			RequestType: c.RequestType,
+			DeviceLock:  c.DeviceLock,
+		}
+		return json.Marshal(&x)
+	case "ClearPasscode":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*ClearPasscode
+		}{
+			RequestType:   c.RequestType,
+			ClearPasscode: c.ClearPasscode,
+		}
+		return json.Marshal(&x)
+	case "EraseDevice":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*EraseDevice
+		}{
+			RequestType: c.RequestType,
+			EraseDevice: c.EraseDevice,
+		}
+		return json.Marshal(&x)
+	case "RequestMirroring":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*RequestMirroring
+		}{
+			RequestType:      c.RequestType,
+			RequestMirroring: c.RequestMirroring,
+		}
+		return json.Marshal(&x)
+	case "Restrictions":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*Restrictions
+		}{
+			RequestType:  c.RequestType,
+			Restrictions: c.Restrictions,
+		}
+		return json.Marshal(&x)
+	case "UnlockUserAccount":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*UnlockUserAccount
+		}{
+			RequestType:       c.RequestType,
+			UnlockUserAccount: c.UnlockUserAccount,
+		}
+		return json.Marshal(&x)
+	case "DeleteUser":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*DeleteUser
+		}{
+			RequestType: c.RequestType,
+			DeleteUser:  c.DeleteUser,
+		}
+		return json.Marshal(&x)
+	case "EnableLostMode":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*EnableLostMode
+		}{
+			RequestType:    c.RequestType,
+			EnableLostMode: c.EnableLostMode,
+		}
+		return json.Marshal(&x)
+	case "InstallApplication":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*InstallApplication
+		}{
+			RequestType:        c.RequestType,
+			InstallApplication: c.InstallApplication,
+		}
+		return json.Marshal(&x)
+	case "AccountConfiguration":
+		var x = struct {
+			RequestType string `json:"request_type"`
+			*AccountConfiguration
+		}{
+			RequestType:          c.RequestType,
+			AccountConfiguration: c.AccountConfiguration,
+		}
+		return json.Marshal(&x)
+	default:
+		return nil, fmt.Errorf("mdm: unknown RequestType: %s", c.RequestType)
+	}
+}
