+++ mdm/mdm/unmarshal_plist.go
@@ -1,0 +1,142 @@
+package mdm
+
+import (
+	"fmt"
+
+	"github.com/pkg/errors"
+)
+
+func (c *Command) UnmarshalPlist(unmarshal func(i interface{}) error) error {
+	var requestType = struct {
+		RequestType string
+	}{}
+	if err := unmarshal(&requestType); err != nil {
+		return errors.Wrap(err, "mdm: unmarshal request type")
+	}
+	c.RequestType = requestType.RequestType
+
+	switch requestType.RequestType {
+	case "ProfileList",
+		"ProvisioningProfileList",
+		"CertificateList",
+		"SecurityInfo",
+		"RestartDevice",
+		"ShutDownDevice",
+		"StopMirroring",
+		"ClearRestrictionsPassword",
+		"UserList",
+		"LogOutUser",
+		"PlayLostModeSound",
+		"DisableLostMode",
+		"DeviceLocation",
+		"TODO_remove":
+		return nil
+	case "InstallProfile":
+		var payload InstallProfile
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.InstallProfile = &payload
+		return nil
+	case "RemoveProfile":
+		var payload RemoveProfile
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.RemoveProfile = &payload
+		return nil
+	case "InstallProvisioningProfile":
+		var payload InstallProvisioningProfile
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.InstallProvisioningProfile = &payload
+		return nil
+	case "RemoveProvisioningProfile":
+		var payload RemoveProvisioningProfile
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.RemoveProvisioningProfile = &payload
+		return nil
+	case "InstalledApplicationList":
+		var payload InstalledApplicationList
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.InstalledApplicationList = &payload
+		return nil
+	case "DeviceInformation":
+		var payload DeviceInformation
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.DeviceInformation = &payload
+		return nil
+	case "DeviceLock":
+		var payload DeviceLock
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.DeviceLock = &payload
+		return nil
+	case "ClearPasscode":
+		var payload ClearPasscode
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.ClearPasscode = &payload
+		return nil
+	case "EraseDevice":
+		var payload EraseDevice
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.EraseDevice = &payload
+		return nil
+	case "RequestMirroring":
+		var payload RequestMirroring
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.RequestMirroring = &payload
+		return nil
+	case "Restrictions":
+		var payload Restrictions
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.Restrictions = &payload
+		return nil
+	case "UnlockUserAccount":
+		var payload UnlockUserAccount
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.UnlockUserAccount = &payload
+		return nil
+	case "DeleteUser":
+		var payload DeleteUser
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.DeleteUser = &payload
+		return nil
+	case "EnableLostMode":
+		var payload EnableLostMode
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.EnableLostMode = &payload
+		return nil
+	case "InstallApplication":
+		var payload InstallApplication
+		if err := unmarshal(&payload); err != nil {
+			return errors.Wrapf(err, "mdm: unmarshal %s command plist", requestType.RequestType)
+		}
+		c.InstallApplication = &payload
+		return nil
+	default:
+		return fmt.Errorf("mdm: unknown RequestType: %s", requestType.RequestType)
+	}
+}
